task 3

+++ 0x10-variadic_functions/3-print_all.c
@@ -31,7 +31,7 @@ void print_all(const char * const format, ...)
 				printf("%s%d", sep, va_arg(args, int));
 				break;
 			case 'f':
-				printf("%s%f", sep, (float)va_arg(args, double));
+				printf("%s%f", sep, va_arg(args, double));
 				break;
 			case 'c':
 				printf("%s%c", sep, va_arg(args, int));
